requirements.txt
openai
faiss-cpu
PyMuPDF
pdfplumber
tabula-py
mysql-connector-python
sqlalchemy
fastapi
uvicorn
pandas
numpy
requests
Pillow